+++ src/main/java/mrriegel/storagenetwork/block/cable/GuiCableProcessing.java
@@ -1,71 +1,24 @@
 package mrriegel.storagenetwork.block.cable;
 
-import java.io.IOException;
-import java.util.List;
 import com.google.common.collect.Lists;
 import mrriegel.storagenetwork.StorageNetwork;
-import mrriegel.storagenetwork.block.AbstractFilterTile;
 import mrriegel.storagenetwork.gui.IPublicGuiContainer;
 import mrriegel.storagenetwork.gui.ItemSlotNetwork;
 import mrriegel.storagenetwork.item.ItemUpgrade;
-import mrriegel.storagenetwork.network.CableDataMessage;
 import mrriegel.storagenetwork.network.CableDataMessage.CableMessageType;
-import mrriegel.storagenetwork.network.CableFilterMessage;
-import mrriegel.storagenetwork.network.CableLimitMessage;
-import mrriegel.storagenetwork.registry.PacketRegistry;
-import mrriegel.storagenetwork.util.UtilTileEntity;
 import mrriegel.storagenetwork.util.data.StackWrapper;
-import net.minecraft.client.gui.FontRenderer;
-import net.minecraft.client.gui.GuiButton;
-import net.minecraft.client.gui.GuiTextField;
-import net.minecraft.client.gui.inventory.GuiContainer;
 import net.minecraft.client.renderer.GlStateManager;
 import net.minecraft.client.resources.I18n;
 import net.minecraft.item.ItemStack;
 import net.minecraft.util.EnumFacing;
-import net.minecraft.util.ResourceLocation;
 import net.minecraft.util.text.TextFormatting;
 import net.minecraftforge.fml.client.config.GuiCheckBox;
 
-public class GuiCableProcessing extends GuiContainer implements IPublicGuiContainer {
-
-  private static final int SQ = 18;
-  private static final int TEXTBOX_WIDTH = 26;
-  private ResourceLocation texture = new ResourceLocation(StorageNetwork.MODID, "textures/gui/cable.png");
-  private GuiCableButton btnPlus, btnMinus, btnWhite, btnOperationToggle, btnImport, btnInputOutputStorage;
-  private TileCable tile;
-  private GuiTextField searchBar;
-  private List<ItemSlotNetwork> itemSlotsGhost;
-  private ItemSlotNetwork operationItemSlot;
-  private GuiCheckBox checkOreBtn;
-  private GuiCheckBox checkMetaBtn;
-  private GuiCableButton pbtnReset;
-  private GuiCableButton pbtnBottomface;
-  private GuiCableButton pbtnTopface;
-  private GuiCableButton buttonRecipe;
-  private GuiCheckBox checkboxNBT;
+public class GuiCableProcessing extends GuiCableBase implements IPublicGuiContainer {
 
   public GuiCableProcessing(ContainerCable inventorySlotsIn) {
     super(inventorySlotsIn);
-    this.xSize = 176;
-    this.ySize = 171;
-    this.tile = inventorySlotsIn.getTile();
-    itemSlotsGhost = Lists.newArrayList();
   }
-
-  @Override
-  public void drawScreen(int mouseX, int mouseY, float partialTicks) {
-    super.drawScreen(mouseX, mouseY, partialTicks);
-    super.renderHoveredToolTip(mouseX, mouseY);
-    drawTooltips(mouseX, mouseY);
-  }
-
-  @Override
-  public void drawBackground(int tint) {
-    super.drawBackground(tint);
-  }
-
-  public final static int FONTCOLOR = 4210752;
 
   @Override
   protected void drawGuiContainerBackgroundLayer(float partialTicks, int mouseX, int mouseY) {
@@ -145,62 +98,6 @@ public class GuiCableProcessing extends GuiContainer implements IPublicGuiContai
     if (operationItemSlot != null && tile.getUpgradesOfType(ItemUpgrade.OPERATION) >= 1) {
       operationItemSlot.drawSlot(mouseX, mouseY);
     }
-  }
-
-  private void drawTooltips(int mouseX, int mouseY) {
-    for (ItemSlotNetwork s : itemSlotsGhost) {
-      if (s != null && s.getStack() != null && !s.getStack().isEmpty() && s.isMouseOverSlot(mouseX, mouseY)) {
-        this.renderToolTip(s.getStack(), mouseX, mouseY);
-      }
-    }
-    if (tile.getUpgradesOfType(ItemUpgrade.OPERATION) >= 1) {
-      operationItemSlot.drawTooltip(mouseX, mouseY);
-    }
-    if (pbtnReset != null && pbtnReset.isMouseOver()) {
-      drawHoveringText(Lists.newArrayList(I18n.format("gui.storagenetwork.refresh")), mouseX, mouseY);
-    }
-    if (this.buttonRecipe != null && buttonRecipe.isMouseOver()) {
-      drawHoveringText(Lists.newArrayList(I18n.format("gui.storagenetwork.recipe.tooltip")), mouseX, mouseY);
-    }
-    if (btnImport != null && btnImport.isMouseOver()) {
-      drawHoveringText(Lists.newArrayList(I18n.format("gui.storagenetwork.gui.import")), mouseX, mouseY);
-    }
-    if (btnInputOutputStorage != null && btnInputOutputStorage.isMouseOver()) {
-      drawHoveringText(Lists.newArrayList(I18n.format("gui.storagenetwork.fil.tooltip_" + tile.getWay().toString())), mouseX, mouseY);
-    }
-    //    if (mouseX > guiLeft + 20 && mouseX < guiLeft + 50 && mouseY > guiTop + 2 && mouseY < guiTop + 30) {
-    //      this.drawHoveringText(Lists.newArrayList(I18n.format("gui.storagenetwork.priority")), mouseX, mouseY, fontRenderer);
-    //    }
-    if (btnWhite != null && btnWhite.isMouseOver()) {
-      String s = tile.isWhitelist() ? I18n.format("gui.storagenetwork.gui.whitelist") : I18n.format("gui.storagenetwork.gui.blacklist");
-      this.drawHoveringText(Lists.newArrayList(s), mouseX, mouseY, fontRenderer);
-    }
-    if (btnPlus != null && btnPlus.isMouseOver()) {
-      this.drawHoveringText(Lists.newArrayList(I18n.format("gui.storagenetwork.priority.up")), mouseX, mouseY, fontRenderer);
-    }
-    if (btnMinus != null && btnMinus.isMouseOver()) {
-      this.drawHoveringText(Lists.newArrayList(I18n.format("gui.storagenetwork.priority.down")), mouseX, mouseY, fontRenderer);
-    }
-    if (pbtnTopface != null && pbtnTopface.isMouseOver()) {
-      this.drawHoveringText(Lists.newArrayList(I18n.format("gui.storagenetwork.processing.recipe")), mouseX, mouseY, fontRenderer);
-      //
-    }
-    if (pbtnBottomface != null && pbtnBottomface.isMouseOver()) {
-      this.drawHoveringText(Lists.newArrayList(I18n.format("gui.storagenetwork.processing.extract")), mouseX, mouseY, fontRenderer);
-      //
-    }
-    if (btnOperationToggle != null && btnOperationToggle.isMouseOver()) {
-      String s = I18n.format("gui.storagenetwork.operate.tooltip", I18n.format("gui.storagenetwork.operate.tooltip." + (tile.isMode() ? "more" : "less")), tile.getLimit(), tile.getOperationStack() != null ? tile.getOperationStack().getDisplayName() : "Items");
-      //   String s = I18n.format("gui.storagenetwork.operate.tooltip");
-      this.drawHoveringText(Lists.newArrayList(s), mouseX, mouseY, fontRenderer);
-    }
-  }
-
-  int FONT = 14737632;
-
-  void drawString(String s, int x, int y) {
-    this.drawString(this.fontRenderer, StorageNetwork.lang(s),
-        x, y, FONT);
   }
 
   @Override
@@ -287,108 +184,8 @@ public class GuiCableProcessing extends GuiContainer implements IPublicGuiContai
   }
 
   @Override
-  protected void mouseClicked(int mouseX, int mouseY, int mouseButton) throws IOException {
-    super.mouseClicked(mouseX, mouseY, mouseButton);
-    ItemStack stackCarriedByMouse = mc.player.inventory.getItemStack().copy();
-    if (operationItemSlot != null && operationItemSlot.isMouseOverSlot(mouseX, mouseY) && tile.getUpgradesOfType(ItemUpgrade.OPERATION) >= 1) {
-      tile.setOperationStack(stackCarriedByMouse);
-      operationItemSlot.setStack(stackCarriedByMouse);
-      int num = searchBar.getText().isEmpty() ? 0 : Integer.valueOf(searchBar.getText());
-      PacketRegistry.INSTANCE.sendToServer(new CableLimitMessage(num, tile.getPos(), stackCarriedByMouse));
-      return;
-    }
-    boolean isRightClick = mouseButton == UtilTileEntity.MOUSE_BTN_RIGHT;
-    boolean isLeftClick = mouseButton == UtilTileEntity.MOUSE_BTN_LEFT;
-    boolean isMiddleClick = mouseButton == UtilTileEntity.MOUSE_BTN_MIDDLE_CLICK;
-    for (int i = 0; i < itemSlotsGhost.size(); i++) {
-      ItemSlotNetwork itemSlot = itemSlotsGhost.get(i);
-      if (itemSlot.isMouseOverSlot(mouseX, mouseY)) {
-        ContainerCable container = (ContainerCable) inventorySlots;
-        StackWrapper stackWrapper = container.getTile().getFilter().get(i);
-        boolean doesExistAlready = container.isInFilter(new StackWrapper(stackCarriedByMouse, 1));
-        //diff rules : if i put it in the left, only check the left, and so on
-        if (i < AbstractFilterTile.FILTER_SIZE / 2) {
-          doesExistAlready = container.isInFilter(new StackWrapper(stackCarriedByMouse, 1), 0, AbstractFilterTile.FILTER_SIZE / 2);
-        }
-        else {
-          doesExistAlready = container.isInFilter(new StackWrapper(stackCarriedByMouse, 1), AbstractFilterTile.FILTER_SIZE / 2, AbstractFilterTile.FILTER_SIZE);
-        }
-        if (!stackCarriedByMouse.isEmpty() && !doesExistAlready) {
-          int quantity = (isRightClick) ? 1 : stackCarriedByMouse.getCount();
-          container.getTile().getFilter().put(i, new StackWrapper(stackCarriedByMouse, quantity));
-        }
-        else {
-          if (stackWrapper != null) {
-            if (isLeftClick || stackWrapper.getSize() <= 0) {
-              container.getTile().getFilter().put(i, null);
-              //              stackWrapper.setSize(stackWrapper.getSize() + (isShiftKeyDown() ? 10 : 1));
-            }
-            else if (isRightClick) {
-              stackWrapper.setSize(stackWrapper.getSize() - 1);
-            }
-            else if (isMiddleClick) {
-              stackWrapper.setSize(stackWrapper.getSize() + 1);
-            }
-          }
-        }
-        //        container.slotChanged();
-        PacketRegistry.INSTANCE.sendToServer(new CableFilterMessage(i, tile.getFilter().get(i), tile.getOre(), tile.getMeta(), checkboxNBT.isChecked()));
-        break;
-      }
-    }
-  }
-
-  @Override
-  protected void actionPerformed(GuiButton button) throws IOException {
-    super.actionPerformed(button);
-    if (btnMinus != null && button.id == btnMinus.id) {
-      tile.setPriority(tile.getPriority() - 1);
-      PacketRegistry.INSTANCE.sendToServer(new CableDataMessage(button.id, tile.getPos()));
-    }
-    else if (btnPlus != null && button.id == btnPlus.id) {
-      tile.setPriority(tile.getPriority() + 1);
-      PacketRegistry.INSTANCE.sendToServer(new CableDataMessage(button.id, tile.getPos()));
-    }
-    else if (btnWhite != null && button.id == btnWhite.id) {
-      tile.setWhite(!tile.isWhitelist());
-      PacketRegistry.INSTANCE.sendToServer(new CableDataMessage(button.id, tile.getPos()));
-    }
-    else if (pbtnTopface != null && button.id == pbtnTopface.id) {
-      int newFace = (tile.getFacingTopRow().ordinal() + 1) % EnumFacing.values().length;
-      tile.processingTop = EnumFacing.values()[newFace];
-      PacketRegistry.INSTANCE.sendToServer(new CableDataMessage(button.id, tile.getPos(), newFace));
-    }
-    else if (pbtnBottomface != null && button.id == pbtnBottomface.id) {
-      //
-      int newFace = (tile.getFacingBottomRow().ordinal() + 1) % EnumFacing.values().length;
-      tile.processingBottom = EnumFacing.values()[newFace];
-      PacketRegistry.INSTANCE.sendToServer(new CableDataMessage(button.id, tile.getPos(), newFace));
-    }
-    else if (btnOperationToggle != null && button.id == btnOperationToggle.id) {
-      if (tile instanceof TileCable) tile.setMode(!tile.isMode());
-      PacketRegistry.INSTANCE.sendToServer(new CableDataMessage(button.id, tile.getPos()));
-    }
-    else if (checkMetaBtn != null && checkOreBtn != null && (button.id == checkMetaBtn.id || button.id == checkOreBtn.id)) {
-      PacketRegistry.INSTANCE.sendToServer(new CableFilterMessage(-1, null, checkOreBtn.isChecked(), checkMetaBtn.isChecked(), this.checkboxNBT.isChecked()));
-    }
-    else {
-      PacketRegistry.INSTANCE.sendToServer(new CableDataMessage(button.id, tile.getPos()));
-    }
-  }
-
-  @Override
-  public void onGuiClosed() {
-    super.onGuiClosed();
-  }
-
-  @Override
   public void drawGradientRectP(int left, int top, int right, int bottom, int startColor, int endColor) {
     super.drawGradientRect(left, top, right, bottom, startColor, endColor);
-  }
-
-  @Override
-  public FontRenderer getFont() {
-    return this.fontRenderer;
   }
 
   @Override
