+++ update.json
@@ -60,6 +60,6 @@
   	,"1.5.3":"Processing cable now allows you to put the same item on both the left and right side.  Added NBT checkbox for filtering items in cables.  Fixed a game crash where network tried to get ore-dictionary data on empty item stacks.  Fix the game sometimes crashing if JEI is not installed.  Fixed a bug where shift-crafting multiple stacks in the remote could consume too many items.  "
 	,"1.5.4":"Fixed Storage Remote sometimes not opening and getting locked up.  "
 	,"1.5.5":"Processing cable now indicates if its recipe is invalid to the player.  Updated priority tooltip.  Fixed delay in processing cables (it no longer looks for impossible speed upgrades).  Update to recipe handler transfer from JEI to avoid getting mismatched NBT items (such as tools or copper cables). "
-  	,"1.5.6":"Fixed Controller not listing correct block name for some machines (Mekanism, exu2).  New config S:BlacklistBlocks added to disable connections to any block, default value contains extrautils2:playerchest.  Removed a debug button that did nothing. "
+  	,"1.5.6":"Fixed Controller not listing correct block name for some machines (Mekanism, exu2).  New config S:BlacklistBlocks added to disable connections to any block, default value contains extrautils2:playerchest.  Removed a debug button that did nothing.  Added scroll-wheel support to certain filter and recipe slots in network cable inventories.  "
   }
 }
